Ajou possibilié de gestion d'une exception

+++ api-safetynet/src/main/java/com/openclassrooms/apisafetynet/controller/PersonController.java
@@ -1,5 +1,6 @@
 package com.openclassrooms.apisafetynet.controller;
 
+import java.net.URI;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +8,7 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.query.Param;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,7 +17,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import com.openclassrooms.apisafetynet.exceptions.UnfindablePersonException;
 import com.openclassrooms.apisafetynet.model.ChildAlertAndFamily;
 import com.openclassrooms.apisafetynet.model.MedicalRecord;
 import com.openclassrooms.apisafetynet.model.Person;
@@ -38,22 +42,49 @@ public class PersonController {
 
     @DeleteMapping("/person/{id_bd}")
     public void deletePerson(@PathVariable("id_bd") String idDb) {
+    	// Possibilité 1 de gérer une exception
+    	Optional<Person> person = personService.getPerson(idDb);
+    	if(person == null ) throw new UnfindablePersonException("La personne avec l'id " + idDb + " est INTROUVABLE.");
         personService.deletePerson(idDb);
     }
     
     @GetMapping("/persons")
     public Iterable<Person> getPersons() {
         return personService.getPersons();
+        
     }
     
     @PostMapping("/person")
-    public Person createPerson(@RequestBody Person person) {
-        return personService.savePerson(person);
+    public ResponseEntity<Void> createPerson(@RequestBody Person person) {
+        Person p = personService.savePerson(person);
+       // Possibilité 2 de gérer une exception
+        if (p == null)
+            return ResponseEntity.noContent().build();
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(person.getId())
+                .toUri();
+        
+        return ResponseEntity.created(location).build();
     }
+    /*
     @PostMapping("/persons")
 	public Iterable<Person> createPerson(@RequestBody Iterable<Person> persons) {
 		return personService.savePersons(persons);
 	}
+    */
+    // trouver quelle response entity renvoyer pour plusieurs enregsiremens 
+    @PostMapping("/persons")
+    public ResponseEntity<Void> createPersons(@RequestBody Iterable<Person> persons) {
+    	Iterable<Person> p = personService.savePersons(persons);
+        if (p == null)
+            return ResponseEntity.noContent().build();
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest().path("packdeperson").build(false).toUri();
+        
+        return ResponseEntity.created(location).build();
+    }
     
     @PutMapping("/person/{id_db}")
     public Person updatePerson(@PathVariable("id_db") String idDb, @RequestBody Person person) {
